requirements.txt
nicegui
mysql-connector-python
yagmail
python-dotenv